Added role for gluster installation.
This change introduces a role for installing RHGS
on all RHHI nodes via gdeploy. The installation
only happens if the `gluster` command is not present
on the target system. If the installation succeeds or
if `gluster` is already present, then the playbook
will iterate over all other nodes in inventory and
verify that they can be probed.

The defaults in `gluster-config/defaults/main.yml`
were also modified to fit a realistic default
installation scenario.