Functioning hosted engine installation.
This change makes it possible for the hosted engine
installation to work. A few variables were removed in
favor of more static configuration. The configuration, at this time,
assumes that the end user has provisoned machines sufficiently
heavy on memory, CPU and disk space that are needed to make
the install succeed.